made conversions explicit, except for some who would fuck up the norm and are irrelevant

+++ bonus/cylinder_bonus.c
@@ -6,7 +6,7 @@
 /*   By: mgroen <mgroen@student.codam.nl>             +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/03/01 12:37:15 by mgroen        #+#    #+#                 */
-/*   Updated: 2022/09/08 14:24:03 by bhoitzin      ########   odam.nl         */
+/*   Updated: 2022/09/11 14:46:40 by bhoitzin      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -98,8 +98,8 @@ double	find_hit_cy(t_scene *scene, t_ray *ray, int *num, int cap)
 	t_ray	*new_ray;
 	double	ret;
 
-	new_ray = malloc(sizeof(t_ray) * scene->amount[CYLINDER]);
-	t = malloc(sizeof(double) * scene->amount[CYLINDER]);
+	new_ray = malloc(sizeof(t_ray) * (unsigned long)scene->amount[CYLINDER]);
+	t = malloc(sizeof(double) * (unsigned long)scene->amount[CYLINDER]);
 	set_values(&ret, num, t);
 	while (*num < scene->amount[CYLINDER])
 	{
